requirements.txt
langchain-community
langchain-core
langchain-huggingface
langchain-groq
python-dotenv